#49 refactor(seat): 공연장 전체 좌석 조회 API - Service 패키지 구조 개선 및 비즈니스 로직 주석 추가
