+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/controller/stadium/StadiumController.java
@@ -12,18 +12,18 @@ import org.springframework.web.bind.annotation.RestController;
 import lombok.RequiredArgsConstructor;
 import wisoft.nextframe.schedulereservationticketing.common.response.ApiResponse;
 import wisoft.nextframe.schedulereservationticketing.dto.seat.seatdefinition.SeatDefinitionListResponse;
-import wisoft.nextframe.schedulereservationticketing.service.stadium.StadiumService;
+import wisoft.nextframe.schedulereservationticketing.service.seat.SeatDefinitionService;
 
 @RestController
 @RequestMapping("/api/v1/stadiums")
 @RequiredArgsConstructor
 public class StadiumController {
 
-	private final StadiumService stadiumService;
+	private final SeatDefinitionService seatDefinitionService;
 
 	@GetMapping("/{id}/seat-definitions")
 	public ResponseEntity<ApiResponse<?>> getSeatDefinitions(@PathVariable UUID id) {
-		final SeatDefinitionListResponse data = stadiumService.getSeatDefinitions(id);
+		final SeatDefinitionListResponse data = seatDefinitionService.getSeatDefinitions(id);
 
 		final ApiResponse<SeatDefinitionListResponse> response = ApiResponse.success(data);
