+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/service/seat/SeatDefinitionService.java
@@ -1,4 +1,4 @@
-package wisoft.nextframe.schedulereservationticketing.service.stadium;
+package wisoft.nextframe.schedulereservationticketing.service.seat;
 
 import java.util.List;
 import java.util.UUID;
@@ -15,7 +15,7 @@ import wisoft.nextframe.schedulereservationticketing.repository.stadium.StadiumR
 
 @Service
 @RequiredArgsConstructor
-public class StadiumService {
+public class SeatDefinitionService {
 
 	private final StadiumRepository stadiumRepository;
 	private final SeatDefinitionRepository seatDefinitionRepository;
@@ -30,9 +30,10 @@ public class StadiumService {
 	@Transactional(readOnly = true)
 	public SeatDefinitionListResponse getSeatDefinitions(UUID stadiumId) {
 		// 1. 공연장 존재 여부 확인
-		stadiumRepository.findById(stadiumId).orElseThrow(EntityNotFoundException::new);
+		stadiumRepository.findById(stadiumId)
+			.orElseThrow(() -> new EntityNotFoundException("해당 공연장을 찾을 수 없습니다."));
 
-		// 2. 공연장의 좌석 데이터 조회
+		// 2. 공연장의 좌석 정보 조회
 		final List<SeatDefinition> seatDefinitions = seatDefinitionRepository.findAllByStadiumIdWithSorting(stadiumId);
 
 		// 3. 조회된 엔티티 목록을 응답 DTO로 변환하여 반환
